backend/requirements.txt
pillow
scikit-learn
numpy
tensorflow
opencv-python-headless
matplotlib
